requirements.txt
fastapi
google-api-python-client
uvicorn
pymongo
discord.py
python-decouple